add python-compat for obj-to-str
