+++ src/dynamic_graph/CMakeLists.txt
@@ -11,6 +11,7 @@ ADD_LIBRARY(${PYTHON_MODULE}
     exception-python.cc
     factory-py.cc
     pool-py.cc
+    python-compat.cc
     signal-base-py.cc
     signal-caster-py.cc
     signal-wrapper.cc
