+++ src/dynamic_graph/entity-py.cc
@@ -1,6 +1,5 @@
 // Copyright 2010, Florent Lamiraux, Thomas Moulard, LAAS-CNRS.
 
-#include <Python.h>
 #include <iostream>
 
 #include <dynamic-graph/entity.h>
